# follow the ICU User Guide recommendation to optimize the perf of InvariantCultureIgnoreCase on Linux