Fixes https://github.com/dotnet/corefx/issues/37691

I have profiled the benchmarks attached in https://github.com/dotnet/corefx/issues/37691 and figured out that `ucol_getSortKey()` slows `InvariantCultureIgnoreCase` string comparison on Linux.

http://icu-project.org/apiref/icu4c/ucol_8h.html#a1f83f9c96a0950e2c22bd5c5c31ff6bf says:

> Note that sort keys are often less efficient than simply doing comparison. For more details, see the ICU User Guide.

While [ICU User Guide](ftp://ftp.software.ibm.com/software/globalization/icu/3.2/icu-3.2-userguide.pdf):

> ucol_getSortKey() can operate in 'preflighting' mode, which returns the amount ofmemory needed to store the resulting sort key. This mode is automatically activated if theoutput buffer size passed is set to zero. Should the sort key become longer than the bufferprovided, function again slips into preflighting mode. The overall performance is poorerthan if the function is called with a zero output buffer . If the size of the sort key returnedis greater than the size of the buffer provided, the content of the result buffer is undefined.In that case, the result buffer could be reallocated to its proper size and the sort keygenerator function can be used again. 
The best way to generate a series of sort keys is to do the following:
1.Create a big temporary buffer on the stack. Typically, this buffer is allocated onlyonce, and reused with every sort key generated. There is no need to keep it as small aspossible. A recommended size for the temporary buffer is four times the length of thelongest string processed.
2.Start the loop. Call ucol_getSortKey()to find out how big the sort key buffer shouldbe, and fill in the temporary buffer at the same time.
3.If the temporary buffer is too small, allocate or reallocate more space for in anoverflow buffer to handle the overflow situations. Fill in the sort key values in theoverflow buffer.
4.Allocate the sort key buffer with the size returned by ucol_getSortKey() and callmemcpy to copy the sort key content from the temp buffer to the sort key buffer. 
5.Loop back to step 1 until you are done.
6.Delete the overflow buffer if you created one.

So I just applied it to the code.

@tarekgh @GrabYourPitchforks PTAL

I have to leave home now, I am going to provide the updated perf numbers tomorrow. Please don't merge until then.