@adamsitnik 2019-06-02
Perf number for benchmarks added in https://github.com/dotnet/performance/pull/535 to the performance repo:

Before:

 |                   Method |    Count |                 Comparison |                Mean | Allocated Memory/Op |
 |------------------------- |--------- | -------------------------- |--------------------:|--------------------:|
 |        GetStringHashCode |       10 | InvariantCultureIgnoreCase |         1,454.43 ns |                   - |
 |        GetStringHashCode |       10 |          OrdinalIgnoreCase |           118.59 ns |                   - |
 |        GetStringHashCode | 10000000 | InvariantCultureIgnoreCase | 1,332,108,533.33 ns |          36377600 B |
 |        GetStringHashCode | 10000000 |          OrdinalIgnoreCase |    91,151,471.67 ns |          20000024 B |

After:

 |                   Method |    Count |                 Comparison |              Mean | Allocated Memory/Op |
 |------------------------- |--------- | -------------------------- |------------------:|--------------------:|
 |        GetStringHashCode |       10 | InvariantCultureIgnoreCase |         756.84 ns |                   - |
 |        GetStringHashCode |       10 |          OrdinalIgnoreCase |         120.80 ns |                   - |
 |        GetStringHashCode | 10000000 | InvariantCultureIgnoreCase | 694,574,080.00 ns |          40000024 B |
 |        GetStringHashCode | 10000000 |          OrdinalIgnoreCase |  91,100,126.79 ns |          20000024 B |

It's two times faster now. 

Why it allocates more memory for getting the hashcode of string of 10000000 characters? The recommended "guessing" size is `length x 4`, but the actual is typically slightly less (3.6, 3.7 or so). Despite that, it's two times faster.