@stephentoub 2019-06-02
> Despite that, it's two times faster.

In a serialized benchmark. The impact of more memory allocated generally didn't show up in such microbenchmarks.