@adamsitnik 2019-06-02
> In a serialized benchmark. The impact of more memory allocated generally doesn't show up in such microbenchmarks.

I agree, however, I believe that such a scenario where we have a very long string and we need the hash of it is very rare. Typically the values in dictionary are long, not the keys.

The ASP.NET app provided by @sebastienros in https://github.com/dotnet/corefx/issues/37691#issuecomment-493143516 uses keys shorter than 10 characters.

I am open to investigating it more if somebody can provide some real world example.