@stephentoub 2019-06-02
> however

I'm not disagreeing with the change; I'm disagreeing with what I perceived as the implications of the comment.  Saying "despite that, it's two times faster" suggests that the extra memory consumption doesn't matter, at least the way I hear the statement.  I just want to make sure we're thinking about such changes holistically; microbenchmarks are a great tool, but they're fallible when it comes to real-world impact.