@stephentoub 2019-06-02
That said, allocating this amount of memory for long strings (I assume it's coming from trying to rent an array from the pool and the pool not storing arrays that long) in order to get a hash code is problematic.  That was the case prior to your change, but it's even more so after.  I am concerned about it.