@adamsitnik 2019-06-02
I see, please excuse me for such simplification.

What I wanted to say was that calling `ucol_getSortKey` two times for very big strings takes much more time than calling it once and allocating 10% more memory.