@adamsitnik 2019-06-02
@jkotas @stephentoub after setting the max guessed size to 1MB we get the x2 perf boost for the strings that are shorter than 1024 * 1024 / 4 and same performance for bigger strings

BEFORE:

 |            Method |   Count |                 Comparison |           Mean | Allocated Memory/Op |
 |------------------ |-------- | -------------------------- |---------------:|--------------------:|
 | GetStringHashCode |      10 | InvariantCultureIgnoreCase |       1.447 us |                   - |
 | GetStringHashCode |  262144 | InvariantCultureIgnoreCase |  34,892.750 us |                   - |
 | GetStringHashCode | 1048576 | InvariantCultureIgnoreCase | 142,401.211 us |           3815328 B |

AFTER:

|            Method |   Count |                 Comparison |             Mean | Allocated Memory/Op |
|------------------ |-------- | -------------------------- |-----------------:|--------------------:|
| GetStringHashCode |      10 | InvariantCultureIgnoreCase |         760.6 ns |                   - |
| GetStringHashCode |  262144 | InvariantCultureIgnoreCase |  17,951,277.1 ns |                   - |
| GetStringHashCode | 1048576 | InvariantCultureIgnoreCase | 141,698,425.0 ns |           3815328 B |