@jkotas 2019-06-02
The Windows implementation of the globalization-aware hash code has the same performance problem (calling the OS API twice). Do we want to apply the same fix there as well?