@jkotas 2019-06-02
dotnet/corefx#37691 also mentions scalability problems (that the slowdown goes from 2x to 6x when there are concurrent calls). Have you looked into that? Maybe [this lock](https://github.com/dotnet/coreclr/blob/master/src/corefx/System.Globalization.Native/pal_collation.c#L402) is a problem - can we switch to lock-free scheme there?