@tarekgh 2019-06-02
Thanks @adamsitnik for optimizing this scenario. I hope you can look at what @jkotas suggested as I am expecting there will be more opportunities to optimize more. let me know if you need any help from me.