@adamsitnik 2019-06-03
> The Windows implementation of the globalization-aware hash code has the same performance problem (calling the OS API twice). Do we want to apply the same fix there as well?

I was thinking about it, but if my understanding of the [docs](https://docs.microsoft.com/en-us/windows/desktop/api/winnls/nf-winnls-lcmapstringex) is correct, the LCMapStringEx does not return actualKeyLength if the guessed size is incorrect but an error instead. So we could get 1 method call for the happy path (perf boost) and 3 method calls for the unhappy path (perf regression). 

```cs
int guessedSize = input.Length * 4;
byte[] buffer = GetBuffer(guessedSize);
int guessedSizeResult = LCMapStringEx(guessedSize, buffer);

if (guessedSizeResult != 0)
{
  // happy path, buffer contains the actual sort key
}
else
{
   int sortKeyLength = LCMapStringEx(0, null); // don't try to guess, ask
   buffer = GetBuffer(sortKeyLength);
   LCMapStringEx(sortKeyLength, buffer); // use returned key length
}
```

Having said that, we could still otpimize the method on Windows if we invest enough time for testing and measuring.

1. Try to guess only for short strings (stackalloc, arraypool.shared.rent  without allocation). Fall back to unhappy path for very rare cases.
2. Use existing code path for longer strings.

> Returns the number of characters or bytes in the translated string or sort key, including a terminating null character, if successful. If the function succeeds and the value of cchDest is 0, the return value is the size of the buffer required to hold the translated string or sort key, including a terminating null character if the input was null terminated.
This function returns 0 if it does not succeed. To get extended error information, the application can call GetLastError, which can return one of the following error codes:
    ERROR_INSUFFICIENT_BUFFER. A supplied buffer size was not large enough, or it was incorrectly set to NULL.