@adamsitnik 2019-06-03
> also mentions scalability problems (that the slowdown goes from 2x to 6x when there are concurrent calls). Have you looked into that? Maybe this lock is a problem - can we switch to lock-free scheme there?

> I hope you can look at what @jkotas suggested as I am expecting there will be more opportunities to optimize more

I am looking into that and experimeting with ThreadStatic with no lock. If I get some visible improvement I am going to send another PR (it would be a bigger change and I would prefere to isolate it)