@adamsitnik 2019-06-04
/azp run